feat: cc-0101-refactor-courts-codes
